# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/albums": {
      "get": {
        "operationId": "getAllAlbums",
        "parameters": [
          {
            "name": "assetId",
            "required": false,
            "in": "query",
            "description": "Only returns albums that contain the asset\nIgnores the shared parameter\nundefined: get all albums",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "shared",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      },
      "post": {
        "operationId": "createAlbum",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/albums/statistics": {
      "get": {
        "operationId": "getAlbumStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumStatisticsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/albums/{id}": {
      "delete": {
        "operationId": "deleteAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      },
      "get": {
        "operationId": "getAlbumInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withoutAssets",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      },
      "patch": {
        "operationId": "updateAlbumInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/albums/{id}/assets": {
      "delete": {
        "operationId": "removeAssetFromAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkIdsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BulkIdResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      },
      "put": {
        "operationId": "addAssetsToAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkIdsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BulkIdResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/albums/{id}/user/{userId}": {
      "delete": {
        "operationId": "removeUserFromAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      },
      "put": {
        "operationId": "updateAlbumUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/albums/{id}/users": {
      "put": {
        "operationId": "addUsersToAlbum",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUsersDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.119.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AddUsersDto": {
        "properties": {
          "albumUsers": {
            "items": {
              "$ref": "#/components/schemas/AlbumUserAddDto"
            },
            "type": "array"
          }
        },
        "required": [
          "albumUsers"
        ],
        "type": "object"
      },
...
      "AlbumResponseDto": {
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumThumbnailAssetId": {
            "nullable": true,
            "type": "string"
          },
          "albumUsers": {
            "items": {
              "$ref": "#/components/schemas/AlbumUserResponseDto"
            },
            "type": "array"
          },
          "assetCount": {
            "type": "integer"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "endDate": {
            "format": "date-time",
            "type": "string"
          },
          "hasSharedLink": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isActivityEnabled": {
            "type": "boolean"
          },
          "lastModifiedAssetTimestamp": {
            "format": "date-time",
            "type": "string"
          },
          "order": {
            "$ref": "#/components/schemas/AssetOrder"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "published": {
            "type": "boolean"
          },
          "shared": {
            "type": "boolean"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "albumName",
          "albumThumbnailAssetId",
          "albumUsers",
          "assetCount",
          "assets",
          "createdAt",
          "description",
          "hasSharedLink",
          "id",
          "isActivityEnabled",
          "owner",
          "ownerId",
          "published",
          "shared",
          "updatedAt"
        ],
        "type": "object"
      },
      "AlbumStatisticsResponseDto": {
        "properties": {
          "notShared": {
            "type": "integer"
          },
          "owned": {
            "type": "integer"
          },
          "shared": {
            "type": "integer"
          }
        },
        "required": [
          "notShared",
          "owned",
          "shared"
        ],
        "type": "object"
      },
      "AlbumUserAddDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AlbumUserRole"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "userId"
        ],
        "type": "object"
      },
      "AlbumUserCreateDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AlbumUserRole"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "role",
          "userId"
        ],
        "type": "object"
      },
      "AlbumUserResponseDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AlbumUserRole"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "role",
          "user"
        ],
        "type": "object"
      },
      "AlbumUserRole": {
        "enum": [
          "editor",
          "viewer"
        ],
        "type": "string"
      },
...
      "AssetFaceWithoutPersonResponseDto": {
        "properties": {
          "boundingBoxX1": {
            "type": "integer"
          },
          "boundingBoxX2": {
            "type": "integer"
          },
          "boundingBoxY1": {
            "type": "integer"
          },
          "boundingBoxY2": {
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "imageHeight": {
            "type": "integer"
          },
          "imageWidth": {
            "type": "integer"
          },
          "sourceType": {
            "$ref": "#/components/schemas/SourceType"
          }
        },
        "required": [
          "boundingBoxX1",
          "boundingBoxX2",
          "boundingBoxY1",
          "boundingBoxY2",
          "id",
          "imageHeight",
          "imageWidth"
        ],
        "type": "object"
      },
...
      "AssetOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
      "AssetResponseDto": {
        "properties": {
          "checksum": {
            "description": "base64 encoded sha1 hash",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duplicateId": {
            "nullable": true,
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "hasMetadata": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isTrashed": {
            "type": "boolean"
          },
          "libraryId": {
            "deprecated": true,
            "description": "This property was deprecated in v1.106.0",
            "nullable": true,
            "type": "string"
          },
          "livePhotoVideoId": {
            "nullable": true,
            "type": "string"
          },
          "localDateTime": {
            "format": "date-time",
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "originalMimeType": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "ownerId": {
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonWithFacesResponseDto"
            },
            "type": "array"
          },
          "resized": {
            "deprecated": true,
            "description": "This property was deprecated in v1.113.0",
            "type": "boolean"
          },
          "smartInfo": {
            "$ref": "#/components/schemas/SmartInfoResponseDto"
          },
          "stack": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AssetStackResponseDto"
              }
            ],
            "nullable": true
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            },
            "type": "array"
          },
          "thumbhash": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AssetTypeEnum"
          },
          "unassignedFaces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "deviceAssetId",
          "deviceId",
          "duration",
          "fileCreatedAt",
          "fileModifiedAt",
          "hasMetadata",
          "id",
          "isArchived",
          "isFavorite",
          "isOffline",
          "isTrashed",
          "localDateTime",
          "originalFileName",
          "originalPath",
          "ownerId",
          "thumbhash",
          "type",
          "updatedAt"
        ],
        "type": "object"
      },
      "AssetStackResponseDto": {
        "properties": {
          "assetCount": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "primaryAssetId": {
            "type": "string"
          }
        },
        "required": [
          "assetCount",
          "id",
          "primaryAssetId"
        ],
        "type": "object"
      },
...
      "AssetTypeEnum": {
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ],
        "type": "string"
      },
...
      "BulkIdResponseDto": {
        "properties": {
          "error": {
            "enum": [
              "duplicate",
              "no_permission",
              "not_found",
              "unknown"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "success"
        ],
        "type": "object"
      },
      "BulkIdsDto": {
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
...
      "CreateAlbumDto": {
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumUsers": {
            "items": {
              "$ref": "#/components/schemas/AlbumUserCreateDto"
            },
            "type": "array"
          },
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "albumName"
        ],
        "type": "object"
      },
...
      "ExifResponseDto": {
        "properties": {
          "city": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "country": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "dateTimeOriginal": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "exifImageHeight": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exifImageWidth": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "exposureTime": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "fNumber": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "fileSizeInByte": {
            "default": null,
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "focalLength": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "iso": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "latitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "lensModel": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "longitude": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "make": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "model": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "modifyDate": {
            "default": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "orientation": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "projectionType": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "rating": {
            "default": null,
            "nullable": true,
            "type": "number"
          },
          "state": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "timeZone": {
            "default": null,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PersonWithFacesResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "faces": {
            "items": {
              "$ref": "#/components/schemas/AssetFaceWithoutPersonResponseDto"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          },
          "updatedAt": {
            "description": "This property was added in v1.107.0",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "faces",
          "id",
          "isHidden",
          "name",
          "thumbnailPath"
        ],
        "type": "object"
      },
...
      "SmartInfoResponseDto": {
        "properties": {
          "objects": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "SourceType": {
        "enum": [
          "machine-learning",
          "exif"
        ],
        "type": "string"
      },
...
      "TagResponseDto": {
        "properties": {
          "color": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "id",
          "name",
          "updatedAt",
          "value"
        ],
        "type": "object"
      },
...
      "UpdateAlbumDto": {
        "properties": {
          "albumName": {
            "type": "string"
          },
          "albumThumbnailAssetId": {
            "format": "uuid",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActivityEnabled": {
            "type": "boolean"
          },
          "order": {
            "$ref": "#/components/schemas/AssetOrder"
          }
        },
        "type": "object"
      },
      "UpdateAlbumUserDto": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AlbumUserRole"
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserResponseDto": {
        "properties": {
          "avatarColor": {
            "$ref": "#/components/schemas/UserAvatarColor"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "profileChangedAt": {
            "format": "date-time",
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "email",
          "id",
          "name",
          "profileChangedAt",
          "profileImagePath"
        ],
        "type": "object"
      },
...
    }
  }
}